Play around with having a generic key material trait
