+++ crates/bitwarden-crypto/src/keys/mod.rs
@@ -23,4 +23,5 @@ pub use kdf::{
     default_argon2_iterations, default_argon2_memory, default_argon2_parallelism,
     default_pbkdf2_iterations, Kdf,
 };
+mod key_material;
 mod utils;
